pkg: convert opam command filters to dune blangs
This just implements conversion of filters on entire commands using
dune's `when` action. Currently it raises a code error if it a `?` unary
operator is encountered (opam syntax for checking whether a variable is
defined). This may be added in a later change though it's less trivial
to convert than the rest of the opam filter dsl and no packages in
opam-repository currently use this feature.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>