Allow use of ppx in dune
This adds an `(include_preprocessed_sources)` stanza which generates
rules to promote preprocessed version of sources.
It's used in a `ppx/` folder at the project root.
For any library or executable stanza with a `(preprocess (pps ...))`
field in <path>, it will generate promotion rules for preprocessed
versions in `ppx/<path>`.

It also modifies `duneboot.ml` to use `ppx/<path>.ml(i)` instead of
`<path>.ml(i)` if it exists, allowing us to use ppx in development
without making it a build dependency.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>