feature: auto_open field
When vendoring code, it's quite useful to change the names of libraries
to prevent issues with linking. Unfortunately, it requires polluting
one's code with what amounts to undoing the renaming. This feature aims
to automate this tedious aspect. In particular,

```
(library
 (auto_open foo))
```

Should automatically insert `-open Foo` for every explicit users of
[foo]. One example where this will come in handy is removing all the
`module Re = Dune_re` noise in our codebase.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: 815ec147-468a-485e-9970-b6520cb196f6 -->