test(pkg): cycles in opam repo
The solver does not check for cyles. This might lead to an invalid build
plan so something to keep in mind.

We also add a test for a situation where there is a cycle in the repo,
but an alternative valid solution is still available.

In both tests the solver ignores the cycle.

Signed-off-by: Ali Caglayan <alizter@gmail.com>