Solver env in lockdir metadata
This adds a `solver_env` field to the lockdir metadata file to help
users understand why a particular solution was chosen. One possible use
case for this is when a lockdir is checked in, developers can see how
similar their system is to the one used to generate the lockdir and
generate an additional lockdir if necesasry.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>