Add dkml-workflows to TEST_DEPS
When setting up the dune codebase on a new machine it's recommended to
run `make dev-deps` to install development dependencies but you also
need to install the `dkml-workflows` package in order to run (say) `dune
build` in the root of this project. This change adds `dkml-workflows` to
the packages installed with `make dev-deps` to remove the need to
manually install it.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>