Remove flags from solver_env
We want it to be possible for users to choose whether to build test
dependencies at build time rather than at solve time. This means that
when generating a lockdir we must effectively set the opam flag
"with-test" to true and include all test dependencies of the local
package (but not of its dependencies).

For now all solver flags are fixed (`with-test` is true and other flags
are false). In the future we will expose a mechanism to let users
control these flags.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>