engine chore: compute copy_rules after [rules_from_source_dir]
The [copy_rules] field of this record is a bit out of place. It is easy
to compute the copy_rules separately, so we should just do that. Also,
this nicely sets up the next step of putting copy rules directly into
the final list, rather than creating an intermediate list that later
gets concatenated into the main one.

Signed-off-by: Philip White <code@trailingwhite.space>