Reduce allocations while loading rules (#8978)
* engine chore: compute copy_rules after [rules_from_source_dir]

The [copy_rules] field of this record is a bit out of place. It is easy
to compute the copy_rules separately, so we should just do that. Also,
this nicely sets up the next step of putting copy rules directly into
the final list, rather than creating an intermediate list that later
gets concatenated into the main one.

Signed-off-by: Philip White <code@trailingwhite.space>

* engine perf: avoid allocating intermediate list of rules

Signed-off-by: Philip White <code@trailingwhite.space>

* engine perf: avoid allocating record on each iteration of fold

Signed-off-by: Philip White <code@trailingwhite.space>

* tests: promote

Previous commits changed the order in which rules are generated, which
can affect the order things appear in error messages. This seems fine.

Signed-off-by: Philip White <code@trailingwhite.space>

* engine perf: replace refs with two accumulators and use better disjoint function

These changes were made due to feedback from other folks.
- Mutating refs is sometimes worse than allocating due to the write
  barrier. However, we can avoid both by maintaining to accumulators.
- Some of the code was using a less efficient method of checking whether
  two sets are disjoint.

Signed-off-by: Philip White <code@trailingwhite.space>

* chore: add [disjoint] into the Set module

It's a pretty generic function, so we might as well add it to Stdune

Signed-off-by: Philip White <code@trailingwhite.space>

* chore: improve module name

Signed-off-by: Philip White <code@trailingwhite.space>

---------

Signed-off-by: Philip White <code@trailingwhite.space>
Co-authored-by: Philip White <code@trailingwhite.space>