fix(pkg): Check that git repo dir is accessible
Instead of going ahead and running git against something that is not an
accessible directory, this will at least raise an exception.

Are there better ways to solve it? Except for showing a possibly better
error message, not really - if someone has replaced the folder with
something we can't access there is no other option than to fail in some
way or other.

Signed-off-by: Marek Kubica <marek@tarides.com>