fix(pkg): fix downloads on windows
opam invokes the cygwin version of curl which means passing % in a unix
shell is fine. We are however calling the Windows version of curl on
Windows which means that % is not a valid token in a command line
enviornment.

To get around this and make it more cross platform, we quote the
outputted error code on our invocation of curl in with double quotes.
This allows the command to work on Windows and all the platforms it
worked on before.

Signed-off-by: Ali Caglayan <alizter@gmail.com>