pkg: check version constraints for lockfile deps
Identifying lockfile deps now works by searching for a set of
dependencies that both satisfies the dependency formula of the package
and is also satisfied by the package versions present in the lockdir. In
the process some common operations on opam formulae is moved into a
separate module. This is in preparation for a later change that will
depend on some of these operations.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>