refactor(engine): optimize sandbox config storage (#8035)
We store a sandbox configuration set for every single action we create.
For a large build, this can add up.

Our old representation was taking 6 words to represent this. This PR
changes it to use only a single word. Moreover, this new set now opaque
to the GC, speeds up comparison, hashing, etc.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>