fix(pkg): Avoid IO when determining the default branch
This moves the default branch determination to when the remote is first
added: it determines the HEAD branch and uses this sole branch as a
tracking branch.

When loading a repo, it determines the default branch from the list of
tracking branches (of which there is now only one, thus it must be the
default).

Signed-off-by: Marek Kubica <marek@tarides.com>