fix(coq): unescape :\ to :
coqdep outputs escaped paths for makefile which means on Windows that
`C:\foo\bar` gets escaped to `C\:\\foo\\bar` causing Dune to interpret
escaped absolute Windows directories as relative ones.

This patch searches for `:\` in the deps output of coqdep and replaces it
with `:` allowing the paths to be interpreted correctly.

Signed-off-by: Ali Caglayan <alizter@gmail.com>