Add delay to address flaky fsevent test
The test checks that exceptions raised in in event handlers cause the
watcher to terminate and the exception is passed to the thread waiting
on the watcher. The test is flaky because the watcher is stopped
directly after some period of time and there's a race between this and
the exception being raised. This change delays stopping the watcher
directly so it's practically guaranteed that there will be time for the
exception to be raised.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>