pkg: substitute custom solver vars in lockfiles
Previsouly only "with-test" and "with-doc" were substituted. This
generalizes partial evaluation of filters to include all solver vars. To
allow users to set "with-doc" but to maintain the current behaviour
where it defaults to "false", it's added to the solver env by default
with value "false".

Also added a case to the logic for simplifying slang expressions so the
generated code is more readable as previously it would generate
expressions like `(when true ...)` and `(when false ...)` when a filter
is applied to individual terms in a command.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>