[ci] Test Coq 8.18
This is an experiment to see if we should test in CI a lower and an
upper bound for Coq.

Note that starting with Coq 8.17, Coq has been split into `coq-core`
and `coq-stdlib`, with `coq-stdlib` being the heavy package.

Unfortunately we still need to install `coq-stdlib` for our tests, as
we need to check that stdlib / installed theories are working
properly, but we could at some point just have a very light
`coq-prelude` package that would take 2 seconds to build and would
work for us too.

Signed-off-by: Emilio Jesus Gallego Arias <e+git@x80.org>