test: fix carriage return from being outputted during build (#9374)
When we don't clear the status line, it gets cleared when we exit. This
causes a carriage return to be output by the inline test runner. When
running the tests in the repo, this would show up as a random space.

After investigating, I isolated the culprit to a recently added test not
having a Status_line.clear at the end. Adding one fixes the original
issue. I also updated the documentation in the tests to explain why we
need to have one.

Signed-off-by: Ali Caglayan <alizter@gmail.com>