fix(pkg): Avoid `git` translating its UI.
Turns out that some Git commands translate their messages e.g. when
running with a `fr_FR` locale. Then the regex fails and as such the code
fails.

This PR makes sure to always run Git in English, though a better
solution would be to find a lower-level command that determines the same
info without having to parse high-level commands that get translated.

Signed-off-by: Marek Kubica <marek@tarides.com>