Remove dependencies on Dune_engine from Targets
This is a pure refactoring. The main goal is to be able to move Dune_engine.Targets into a separate package which can be used from dune_cache. However, I believe that it also makes the code simpler to follow overall.

The main change is that we move everything to do with async file operations into Build_system which really is the only user of those, anyway.

In addition, we also slightly generalise [Targets.Produced.collect_digests] and remove an unused pretty-printing function.

Signed-off-by: Roman Leshchinskiy <rleshchinskiy@janestreet.com>