[coq] Allow to disable Coq language deprecation warning.
I had to adapt the warnings infra a little bit, IMHO that is fine, tho
it would be nice to see if we can recover the vendor behavior.

I also corrected the hint message a little bit, as it was a bit
confusing IMHO, but will place that in the next commit as to keep
diffs tidy.

Signed-off-by: Emilio Jesus Gallego Arias <e+git@x80.org>