fix: stop using slang parser in regular dune files (#9530)
This was accidentally used to parse run actions in regular dune files.
This was problematic as the slang dsl is still experimental and intended
for use in package management only at least for now. This change
specializes the parser used for run actions so that the slang dsl parser
is only used when parsing lockfiles.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>