fix(pkg): Read the main branch correctly even if tags exist
The issue was that `git remote show -n <handle>` will return tags as
branches, so if a repo has tags, it will fail to determine the default
branch.

`git remote show` is problematic in many ways (e.g. it localizes output,
thus already having lead to issues), so instead this code reads the
branch from the fetch refspec, which is locale-independent.

This does not fix the issue that submodules are not supported yet, but
it will determine the branch correctly on the way to determining
submodules.

Signed-off-by: Marek Kubica <marek@tarides.com>