fix(boot): make bootstrap reproducible by default
Fixes #9507

In `duneboot.ml`, there are 2 strategies to build `dune.exe`:
- separate compilation (build all modules and link them together in
  several commands)
- single command compilation (pass all modules to the compiler driver)

Separate compilation can be executed in parallel when `-j` is set
accordingly, which speeds up the build.
But when only one job can be executed at once (`-j 1`), it is faster to
use a single command.

So, we would switch to single command build when `-j 1` was passed.
However, this is surprising because it produces a different binary
depending on the number of parallel jobs (see #9507). So this disables
this optimization by default; it can be enabled by hand by passing
`--use-single-command`.

(NB: this does not apply to Windows, which always uses single command
compilation)

Signed-off-by: Etienne Millon <me@emillon.org>