doc: have the 'OCaml flags' documentation point to complete examples
I had forgotten how to disable a warning with Dune, unfortunately the
"OCaml flags" documentation does not actually contain any reusable
code snippet doing this. In fact, the information is in the doc,
hidden in the "Quickstart" and the "FAQ" pages.

This commit adds references from the "OCaml flags" reference section
to those usage examples.

Signed-off-by: Gabriel Scherer <gabriel.scherer@gmail.com>