bench: apply all patches in monorepo benchmark
The monorepo benchmark applies patches to fix packages that don't
normally build in a monorepo setting, and to correct other errors in
packages. The fixes are upstreamed to the individual packages but won't
be visible to the monorepo benchmark until the monorepo is updated
(which we do sparingly). In the meantime we maintain a set of patches.
Until recently the patches were applied while assembling the duniverse
directory containing the monorepo. Reassembling the monorepo can take a
long time, so recently we changed it so that the duniverse directory
contains unpatched packages and patches are applied to a local copy of
the monorepo immediately before running the benchmark. This will allow
us to add patches without needing to regenerate the duniverse directory
which will let us fix package errors faster.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>