feature(coq): extra_sources field for coq.theory stanza
We add an extra_sources field to the coq-theory and coq.extraction
stanzas that allows the decleration of extra sources. This allows Coq
features such as Load and Extra Dependency to be used correctly.

We add two tests that demonstrates how it can be used for those
features.

The extra_sources field finds the declared files and makes sure that
Dune declares a dependency on them before calling coqdep for the theory
being built.

Signed-off-by: Ali Caglayan <alizter@gmail.com>