pkg: enforce opam-compatible package names
This refactors package-management code to use
`Dune_lang.Package_name.Opam_compatible.t` for all package names.
Previously we were checking for opam compatibility in some package
management operations but not others leading to inconsistent behaviour.
This removes the inconsistency by making it impossible to represent an
opam incompatible package name in package management code.

Note that there's still an inconsistency in the treatment of package
names in the rest of dune but we can't remove that currently as people
may depend on it. There's a discussion of this problem at:
https://github.com/ocaml/dune/issues/9371

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>