refactor(pkg): make remote reading a little more resilient
Previously, the algorithm of looking up inside a remote would treat
everything after [remote $name] as belonging to the remote. That's not
correct because we can have other remotes after what we found.

This PR makes sure to first get the section that belongs to a remote and
limits the search to this section.

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: 038d569e-b2d8-49ad-aab9-f1d1fa4b3d44 -->