boot: remove single-command bootstrap
Fixes #9507

The bootstrap process can use 2 strategies:
- parallel: run compile commands in parallel and link the rest
- single-command: run ocamlopt with a long list of arguments

The single-command strategy is used if win32 or if `-j 1` is set
(implicitly or explicitly).

One problem is that: parallel and single-command create different
binaries (#9507).

The assumption was that single-command would be faster than `-j 1` on
Linux and any parallel build on Windows. However, from a quick benchmark
the assumption is true on Linux but false on Windows.

The tiny savings in the case of Linux where only a single core is
available are not enough to justify the extra code path and
reproducibility gotcha.

Signed-off-by: Etienne Millon <me@emillon.org>