pkg: reuse downloaded archives within single run
Archives downloaded over http by package management code will be cached
in a temporary directory that persists during a single invocation of
dune. This is intended to reduce the number of times that dune downloads
the same file during package management operations. Note that it's
common in the ocaml ecosystem for a single source archive to contain
multiple packages.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>