chore: add missing unix dependencies
See #8607 for a recent example.

OCaml 5 adds a `ocaml_deprecated_auto_include` alert when the `unix`
library is referenced but without an explicit dependency.

This commit fixes these alerts by adding a dependency to `unix` in
places that refer to `Unix` directly (ie, it makes the metadata correct
thus removing the alert).

Signed-off-by: Etienne Millon <me@emillon.org>