pkg: fix bug preventing paths from being sandboxed (#9940)
Pform variables which appear in build/install commands that refer to
paths (e.g. `%{target}`) are resolved in two steps. First the normal
pform substitution is applied, replacing the variable with some path,
then the path is further transformed to refer to the corresponding path
within the sandbox specific to the package being built/installed.

Path pform variables which appeared in string interpolations (such as
`--target=%{target}`) were erroneously not subject to the second
transformation, meaning that they were left as the original, unsandboxed
path. This was due to string interpolations being collapsed into strings
too early so that path information was lost at the point when the
sandboxing transformation is applied.

This fixes this by introducing the concept of "deferred concatenations"
which represent a sequence of values to be concatenated in the future.
This lets the path-iness of the path variable components of string
interpolations to be preserved until the point when the sandboxing
transformation is applied.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>