Add quotes around field name in error messages
There are several places where a dune config will use a field named
"name", and if such a field is omitted, prior to this change dune would
print the error 'field name missing' which is ambiguous. This change
puts quotes around the field name (ie. 'field "name" missing') so it's
clear that a field named "name" is missing, and not the name of a field.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>