refactor: high-level dune-package rewriting
`dune install` rewrites `dune-package` files on the fly to store
sections as computed by the installation step.

The previous implementation used to rely on sexp manipulation, which is
brittle and does not evolve with the `dune-package` codecs.

This uses a more end-to-end approach where the data structure is
decoded, modified, and encoded again.

Some aspects are not as easy as they seem because there are different
encodings for paths; some are relative to the package root, other from
the `Stublibs` section, so the API is not as simple as it could be.

Signed-off-by: Etienne Millon <me@emillon.org>