pkg: add depopts to depends when available
When a package's optional dependency is among the packages in the
dependency solution for the project, we want that package's build take
place after its optional dependency has been built. This allows the
package's build commands to observe that its optional dependency is
installed. This can be achieved by adding a package's optional
dependency to its lockfile's `depends` field if the optional dependency
is part of the dependency solution.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>