Fix gh10408 (#10444)
When updating the PATH variable with the `withenv` action, dune would
previously clobber the existing PATH from the environment rather than
updating it. Withenv applies environment updates iteratively starting
from some base environment. Prior to this change, the base environment
did not include the global environment, so as far as withenv knew the
PATH was originally empty. The fix is to start withenv with an
environment which includes the global environment.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>