[3.15] Backport #10446 (#10450)
* Add reproduction test case for #10149.

Signed-off-by: Rodolphe Lepigre <rodolphe@bedrocksystems.com>
Signed-off-by: Ali Caglayan <alizter@gmail.com>

* stop recording deps in Action_builder.contents

PR #9552 changed the semantics of Action_builder.contents so that it
records a dependency. The Coq rules were relying on the previous
behaviour of not recording a dependency in order to generate a build
plan. This caused issue #10149 to appear.

This PR fixes #10149 by removing the dependency being recorded in
Action_builder.contents matching the previous semantics.

Signed-off-by: Ali Caglayan <alizter@gmail.com>

---------

Signed-off-by: Rodolphe Lepigre <rodolphe@bedrocksystems.com>
Signed-off-by: Ali Caglayan <alizter@gmail.com>
Co-authored-by: Rodolphe Lepigre <rodolphe@bedrocksystems.com>
Co-authored-by: Ali Caglayan <alizter@gmail.com>