fix: handle trailing / in Path.External.drop_prefix
In a nutshell:

    # drop_prefix ~prefix:"/a/b/c/" "/a/b/c/d/e"
    Some "d/e"

This improves the test situation on nix (#8203) in the default devshell,
because the paths in `OCAMLPATH` can have trailing slashes.

More precisely, the `--sanitize-for-tests` options uses
`Path.External.drop_prefix` to replace elements of `OCAMLPATH` by the
`FINDLIB` string. If `OCAMLPATH` contains paths with trailing slashes,
these prefixes are not sanitized properly.

Signed-off-by: Etienne Millon <me@emillon.org>