Add experimental toolchain subcommand
This adds a command and module for downloading, building, and
installing the ocaml compiler toolchain inside the user's ~/.cache directory.

For now this works by hardcoding specific supported versions of the
compiler toolchain. In the future we will probably changes this to
read information about toolchain packages from the opam repo.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>