pkg: Undo change reversing order of env updates (#10474)
In 3ea5444cd the order of env updates performed by packages was
reversed. This brought the behaviour of dune closer to opam but
introduced a regression which prevents ocamlfind from building with
dune package management.

Here's an issue for tracking investigating why this change caused the
regression: https://github.com/ocaml/dune/issues/10473

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>