Automatically lock on `build` if there is no lock dir
This is more of an exploration of the toolchains work as it would
automatically lock whenever any user that does not yet have a lock file
calls `dune build`.

For an inclusion in the main dune this would need to be configurable
on... something. The right criterion is up for discussion.

Signed-off-by: Marek Kubica <marek@tarides.com>