fix: virtual libraries bug
When compiling an implementation of a virtual library, there's a check
that makes sure we don't the virtual library doesn't exist in the
closure of the implementation.

This check tried to compute the linking closure of the library to do so.
However, the linking closure might not be complete if the implementation
contains other virtual library.

To fix the issue, we use a "partial" linking closure that tries to
compute the closure as much as possible, but doesn't fail on missing
implementation.

Fix #10460

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: eed8cded-17bc-425f-95b8-3e0337e2881d -->