Allow toolchains to be disabled with flag
Also fixes a bug that was introduced with the initial toolchains
implementation that leads to a memo cycle when resolving the
ocaml-variants package. Breaking the memo cycle also let us remove the
code that tracks the compiler dependency of individual packages which
was adimitedly a hack to work around the same circular dependency that
caused the memo cycle. Removing this code should make it less
contraversial to merge the toolchains branch upstream.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>