Add developer preview features
This adds several experimental features to dune which are disabled by
default by some ad-hoc feature flags. The two notable features are
"toolchains" and "autolocking". These features are included in
experimental versions of dune used in interviews to evaluate dune's
developer experience. The goal of upstreaming these features is to
reduce the maintenance burden of a long-lived feature branch.

Toolchains refers to a feature where dune downloads and builds ocaml
compiler packages using an alternative mechanism to regular dune
package management. Toolchains are installed into
`$XDG_CACHE_DIR/dune/toolchains/<name.version>`. Both the
ocaml-base-compiler and ocaml-variants compilers are supported. In the
case of the latter, multiple different configurations may be installed
concurrently, by concatenating the names of enabled options to the end
of the name of the directory containing the toolchain.

The toolchains feature is a temporary workaround for the fact that the
compiler packages are not currently relocatable. That is, the
artifacts from the compiler packages built inside dune's build sandbox
cannot be used after moving them into the dune cache or their eventual
location within the _build directory. Work is currently being done to
make the compiler relocatable, at which point dune will be able to
build compiler packages in its sandbox and the toolchains feature can
be deleted.

Autolocking refers to generating and updating the lock directory in
any command that would use the metadata it contains (e.g. `dune
build`). This removes the need to explicitly run `dune pkg lock`,
bringing the behavior of dune more in line with that of other popular
language package managers such as cargo and npm.

Co-authored-by: Alpha Issiaga DIALLO <moyodiallo@gmail.com>
Co-authored-by: Etienne Millon <me@emillon.org>
Co-authored-by: Marek Kubica <marek@tarides.com>
Signed-off-by: Stephen Sherratt <stephen@sherra.tt>