Add toolchains feature
Toolchains refers to a feature where dune downloads and builds ocaml
compiler packages using an alternative mechanism to regular dune
package management. Toolchains are installed into
$XDG_CACHE_DIR/dune/toolchains/<name.version>. Both the
ocaml-base-compiler and ocaml-variants compilers are supported. In the
case of the latter, multiple different configurations may be installed
concurrently, by concatenating the names of enabled options to the end
of the name of the directory containing the toolchain.

The toolchains feature is a temporary workaround for the fact that the
compiler packages are not currently relocatable. That is, the
artifacts from the compiler packages built inside dune's build sandbox
cannot be used after moving them into the dune cache or their eventual
location within the _build directory. Work is currently being done to
make the compiler relocatable, at which point dune will be able to
build compiler packages in its sandbox and the toolchains feature can
be deleted.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>