pkg: respect avoid-versions between versions of pkg (#10642)
* pkg: respect avoid-versions between versions of pkg

Dune vendors the opam-0install-solver package and uses it to solve
package dependencies, however opam-0install-solver does not respect
the avoid-versions flag of opam packages. This can lead to situations
where unstable releases of packages (notably of the compiler packages)
are chosen in favor of the latest stable release of a package.

This change implements a partial workaround for this limitation where
the solver will prefer versions of packages that lack the
avoid-version flag.

Note that this does not fully implement the avoid-versions flag. If a
package depends on a disjunction of different packages then the
avoid-versions flag is not considered when choosing which package in
the disjunction to choose.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>

* Rework compare code according to review

Signed-off-by: Marek Kubica <marek@tarides.com>

---------

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>
Signed-off-by: Marek Kubica <marek@tarides.com>
Co-authored-by: Marek Kubica <marek@tarides.com>