chore: use 5.1.1 for tests
This bumps the version of OCaml used in tests to 5.1.1.

The changes are:

- version numbers are changed
- CI is adjusted
- some hashes are changed
- the layout of the compiler is changed
- some error messages are changed
- Interesting changes to note:
  - on OCaml 5, with the dev profile, jsoo triggers a warning
  - the location of cmt files in install files is changed, because they
    are compressed and binary substitution does not happen

At the time this PR is opened, 5.2 is available but requires more
changes around warnings and alerts in alias modules so it will be
updated in a second step.

Also, there are a few components that we're leaving behind:

- the monorepo benchmarks can be updated separately
- the benchmark in bench.Dockerfile can be updated separately
- coq cannot use OCaml 5 at the moment so it's forced to use 4.x
- 5.1.1 doesn't support MSVC so we leave the build on 4.14 for now.

Signed-off-by: Etienne Millon <me@emillon.org>