test: fix oneshot webserver for large binary files
Fixes several issues where the oneshot webserver used for testing
package downloading would exit before the transfer was complete. This
involves transefering data as byte arrays rather than as strings, and
properly shutting down the socket before exiting the server.

The only remaining use case for the `respond_string` function was in
sending 404 errors, so I've replaced it with a specialized function
that only sends 404 errors.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>